+++ src/components/ProductList/BestProductList.jsx
@@ -5,6 +5,7 @@ import { getProducts } from "../../services/api";
 import SectionTitle from "../ui/SectionTitle";
 import ProductListResults from "./ProductListResults";
 import useAsync from "../../hooks/useAsync";
+import ProductListSkeleton from "../ui/Skeletons/ProductListSkeleton";
 
 const ORDER_BY = "favorite";
 const LIST_TYPE = "best";
@@ -35,9 +36,11 @@ const BestProductList = ({ title, pageSize }) => {
   return (
     <div css={ProductListStyle}>
       <SectionTitle title={title} />
+      {isLoading && (
+        <ProductListSkeleton pageSize={pageSize} listType={LIST_TYPE} />
+      )}
+      {loadingError && <p>상품 목록을 가져오지 못했습니다.</p>}
       <ProductListResults
-        isLoading={isLoading}
-        isError={loadingError}
         products={products}
         pageSize={pageSize}
         listType={LIST_TYPE}
