+++ src/components/ProductList/ProductListError.jsx
@@ -1,0 +1,12 @@
+/** @jsxImportSource @emotion/react */
+import { ProductListStyle } from "./ProductList";
+
+const ProductListError = () => {
+  return (
+    <div css={ProductListStyle}>
+      <p>상품 목록을 가져오지 못했습니다.</p>
+    </div>
+  );
+};
+
+export default ProductListError;
